ashesi_market_django/marketplace/urls.py
"""
API URL routing
"""
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from . import views

router = DefaultRouter()
router.register(r'categories', views.CategoryViewSet, basename='category')
router.register(r'products', views.ProductViewSet, basename='product')
router.register(r'orders', views.OrderViewSet, basename='order')
router.register(r'reviews', views.ReviewViewSet, basename='review')

urlpatterns = [
    # CSRF Token
    path('csrf/', views.csrf_token_view, name='csrf-token'),
    
    # Authentication
    path('auth/register/', views.register_view, name='register'),
    path('auth/login/', views.login_view, name='login'),
    path('auth/logout/', views.logout_view, name='logout'),
    path('auth/user/', views.current_user_view, name='current-user'),
    
    # Profile
    path('profile/', views.profile_view, name='profile'),
    path('users/<int:user_id>/', views.user_profile_public, name='user-profile'),
    
    # Cart
    path('cart/', views.cart_view, name='cart'),
    path('cart/add/', views.add_to_cart, name='add-to-cart'),
    path('cart/update/<int:item_id>/', views.update_cart_item, name='update-cart-item'),
    path('cart/remove/<int:item_id>/', views.remove_from_cart, name='remove-from-cart'),
    
    # Checkout
    path('checkout/', views.checkout, name='checkout'),
    
    # Router URLs
    path('', include(router.urls)),
]
