ashesi_market_django/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
Pillow>=10.0.0
python-decouple>=3.8
django-filter>=23.0
# For MySQL on Windows, use PyMySQL instead of mysqlclient
# Uncomment the line below if you want to use MySQL
# PyMySQL>=1.1.0
